+++ Tests/TuistKitTests/Services/GraphServiceTests.swift
@@ -2,6 +2,7 @@ import DOT
 import Foundation
 import GraphViz
 import TSCBasic
+import TuistGraph
 import TuistPlugin
 import TuistSupport
 import XcodeProj
@@ -59,6 +60,36 @@ final class GraphServiceTests: TuistUnitTestCase {
         let expected = "graph { }"
         // Then
         XCTAssertEqual(got, expected)
+        XCTAssertPrinterOutputContains("""
+        Deleting existing graph at \(graphPath.pathString)
+        Graph exported to \(graphPath.pathString)
+        """)
+    }
+    
+    func test_run_whenJson() throws {
+        // Given
+        let temporaryPath = try self.temporaryPath()
+        let graphPath = temporaryPath.appending(component: "graph.json")
+        let projectManifestPath = temporaryPath.appending(component: "Project.swift")
+
+        try FileHandler.shared.touch(graphPath)
+        try FileHandler.shared.touch(projectManifestPath)
+
+        // When
+        try subject.run(
+            format: .json,
+            layoutAlgorithm: .dot,
+            skipTestTargets: false,
+            skipExternalDependencies: false,
+            targetsToFilter: [],
+            path: temporaryPath,
+            outputPath: temporaryPath
+        )
+        let got = try FileHandler.shared.readTextFile(graphPath)
+        
+        let result = try JSONDecoder().decode(GraphOutput.self, from: got.data(using: .utf8)!)
+        // Then
+        XCTAssertEqual(result, GraphOutput.init(name: "graph", path: "/", projects: [:]))
         XCTAssertPrinterOutputContains("""
         Deleting existing graph at \(graphPath.pathString)
         Graph exported to \(graphPath.pathString)